requirements.txt
mutagen
requests
wxpython
pycryptodome